refactor: :art: Move PropsMetadata into its own file
